requirements.txt
numpy
tqdm
torch
torchvision
pillow
einops
wandb